@@ -9,20 +9,23 @@
   "files": [
     "/bin",
     "/dist",
+    "/config.json",
     "/npm-shrinkwrap.json",
     "/oclif.manifest.json"
   ],
   "dependencies": {
     "@coral-xyz/anchor": "^0.27.0",
+    "@lightprotocol/zk.js": "^0.3.2-alpha.8",
     "@oclif/core": "^2",
-    "@oclif/plugin-help": "^5",
-    "@oclif/plugin-plugins": "^2.4.7",
+    "@oclif/plugin-plugins": "^3.0.1",
+    "@solana/spl-token": "^0.3.7",
     "@solana/web3.js": "^1.76.0",
-    "bn.js": "^5.2.1",
-    "circomlibjs": "^0.1.7",
+    "@typescript-eslint/eslint-plugin": "^5.59.7",
+    "@typescript-eslint/parser": "^5.59.7",
+    "cli-spinners": "^2.9.0",
     "dotenv": "^16.0.3",
     "fs": "^0.0.1-security",
-    "light-sdk": "file:../light-sdk-ts"
+    "path": "^0.12.7"
   },
   "devDependencies": {
     "@oclif/test": "^2.3.17",
@@ -31,7 +34,7 @@
     "@types/mocha": "^9.0.0",
     "@types/node": "^16.18.26",
     "chai": "^4",
-    "eslint": "^7.32.0",
+    "eslint": "^8.41.0",
     "eslint-config-oclif": "^4",
     "eslint-config-oclif-typescript": "^1.0.3",
     "mocha": "^9",
