@@ -1,9 +1,26 @@
 import { Args, Command, Flags } from "@oclif/core";
 import { PublicKey } from "@solana/web3.js";
-import { getConnection } from "../../utils";
+import {
+  CustomLoader,
+  generateSolanaTransactionURL,
+  getConnection,
+  setAnchorProvider,
+} from "../../utils/utils";
+import { getOrCreateAssociatedTokenAccount, mintTo } from "@solana/spl-token";
+import {
+  ADMIN_AUTH_KEYPAIR,
+  MINT,
+  airdropSol,
+  convertAndComputeDecimals,
+} from "@lightprotocol/zk.js";
+import { BN } from "@coral-xyz/anchor";
 
 class AirdropCommand extends Command {
-  static description = "Perform an airdrop to a user";
+  static description = "Perform a native Solana or SPL airdrop to a user";
+
+  protected finally(_: Error | undefined): Promise<any> {
+    process.exit();
+  }
 
   static flags = {
     amount: Flags.integer({
@@ -11,14 +28,22 @@ class AirdropCommand extends Command {
       description: "The amount to airdrop",
       required: true,
     }),
+    token: Flags.string({
+      char: "t",
+      description: "The token to airdrop",
+      required: true,
+    }),
   };
 
-  static examples = [`$ light airdrop --amount 2000000000 <userPublicKey>`];
+  static examples = [
+    `$ light airdrop --token SOL --amount 1.0 <userPublicKey>`,
+    `$ light airdrop --token USDC --amount 10 <userPublicKey>`,
+  ];
 
   static args = {
     userPublicKey: Args.string({
       name: "userPublicKey",
-      description: "The public key of the user",
+      description: "The solana public key of the user",
       required: true,
     }),
   };
@@ -27,23 +52,49 @@ class AirdropCommand extends Command {
     const { args, flags } = await this.parse(AirdropCommand);
 
     const { userPublicKey } = args;
-    const { amount } = flags;
+    const { amount, token } = flags;
+
+    const loader = new CustomLoader("Performing the airdrop...");
+    loader.start();
+
+    let response;
 
     try {
-      const connection = await getConnection();
+      const provider = await setAnchorProvider();
 
-      const res = await connection.requestAirdrop(
-        new PublicKey(userPublicKey),
-        amount
-      );
+      if (token.toLowerCase() === "sol") {
+        response = await airdropSol({
+          provider: provider,
+          amount: convertAndComputeDecimals(amount, new BN(1e9)),
+          recipientPublicKey: new PublicKey(userPublicKey),
+        });
+      } else {
+        let tokenAccount = await getOrCreateAssociatedTokenAccount(
+          provider.connection,
+          ADMIN_AUTH_KEYPAIR,
+          MINT,
+          new PublicKey(userPublicKey)
+        );
 
-      await connection.confirmTransaction(res, "confirmed");
+        response = await mintTo(
+          provider.connection,
+          ADMIN_AUTH_KEYPAIR,
+          MINT,
+          tokenAccount.address,
+          new PublicKey(userPublicKey),
+          amount,
+          []
+        );
+      }
 
       this.log(
-        `Airdrop successful for user: ${userPublicKey}, amount: ${amount}`
+        `\nAirdrop successful for user: ${userPublicKey}, amount: ${amount} ${token}`
       );
+      this.log(generateSolanaTransactionURL("tx", response!, "custom"));
+      loader.stop();
     } catch (error) {
-      this.error(`Airdrop failed: ${error}`);
+      loader.stop();
+      this.error(`\nAirdrop failed: ${error}`);
     }
   }
 }
