@@ -1,6 +1,6 @@
 import { Command, Flags } from "@oclif/core";
-import { User, Balance, InboxBalance, Utxo } from "light-sdk";
-import { getUser } from "../../utils";
+import { User, Balance, InboxBalance, Utxo } from "@lightprotocol/zk.js";
+import { CustomLoader, getUser } from "../../utils/utils";
 
 class BalanceCommand extends Command {
   static description =
@@ -10,59 +10,84 @@ class BalanceCommand extends Command {
     balance: Flags.boolean({
       char: "b",
       description: "Retrieve the balance",
-      default: true,
-      exclusive: ["inbox", "utxos"],
+      default: false,
     }),
     inbox: Flags.boolean({
       char: "i",
       description: "Retrieve the inbox balance",
       default: false,
-      exclusive: ["balance", "utxos"],
     }),
     utxos: Flags.boolean({
       char: "u",
       description: "Retrieve the UTXOs",
       default: false,
-      exclusive: ["balance", "inbox"],
+    }),
+    inboxUtxos: Flags.boolean({
+      char: "x",
+      description: "Retrieve the inbox UTXOs",
+      default: false,
     }),
     latest: Flags.boolean({
       char: "l",
-      description: "Retrieve the latest balance/inbox balance/UTXOs",
+      description: "Retrieve the latest balance, inbox balance, or UTXOs",
       default: true,
     }),
   };
 
+  protected finally(_: Error | undefined): Promise<any> {
+    process.exit();
+  }
+
   static examples = [
-    "$ light balance",
+    "$ light balance --balance",
     "$ light balance --inbox",
-    "$ light balance --utxos",
+    "$ light balance --utxos --inbox",
+    "$ light balance --inboxUtxos",
     "$ light balance --latest=false",
   ];
 
   async run() {
     const { flags } = await this.parse(BalanceCommand);
-    const { balance, inbox, utxos, latest } = flags;
+    const { balance, inbox, utxos, latest, inboxUtxos } = flags;
 
-    const user = await getUser();
+    const loader = new CustomLoader("Retrieving balance...");
+
+    loader.start();
+
+    const user: User = await getUser();
 
     try {
       if (balance) {
         const result = await user.getBalance(latest);
         this.logBalance(result);
-      } else if (inbox) {
+      }
+      if (inbox) {
         const result = await user.getUtxoInbox(latest);
         this.logInboxBalance(result);
-      } else if (utxos) {
-        const result = await user.getUtxos(latest);
+      }
+      if (utxos) {
+        const result = await user.getAllUtxos();
         this.logUTXOs(result);
       }
+      if (inboxUtxos) {
+        const result = await user.getUtxoInbox();
+        const utxos: Utxo[] = [];
+        for (const iterator of result.tokenBalances.values()) {
+          iterator.utxos.forEach((value) => {
+            utxos.push(value);
+          });
+        }
+        this.logUTXOs(utxos);
+      }
+      loader.stop();
     } catch (error) {
-      this.error(`Error retrieving balance, inbox balance, or UTXOs ${error}`);
+      loader.stop();
+      this.error(`Error retrieving balance, inbox balance, or UTXOs: ${error}`);
     }
   }
 
   private logBalance(balance: Balance) {
-    this.log("--- Balance ---");
+    this.log("\n--- Balance ---");
     this.log("Token Balances:", balance.tokenBalances);
     this.log("Program Balances:", balance.programBalances);
     this.log("NFT Balances:", balance.nftBalances);
@@ -77,7 +102,7 @@ class BalanceCommand extends Command {
   }
 
   private logInboxBalance(inboxBalance: InboxBalance) {
-    this.log("--- Inbox Balance ---");
+    this.log("\n--- Inbox Balance ---");
     this.log("Token Balances:", inboxBalance.tokenBalances);
     this.log("Program Balances:", inboxBalance.programBalances);
     this.log("NFT Balances:", inboxBalance.nftBalances);
@@ -96,13 +121,13 @@ class BalanceCommand extends Command {
   }
 
   private logUTXOs(utxos: Utxo[]) {
-    this.log("--- UTXOs ---");
+    this.log("\n--- UTXOs ---");
     for (const utxo of utxos) {
       this.log("UTXO:");
-      this.log(`  Amount: ${utxo.amounts}`);
-      this.log(`  Asset: ${utxo.assets}`);
-      this.log(`  Commitment: ${utxo._commitment}`);
-      this.log(`  Index: ${utxo.index}`);
+      this.log(`Amount: ${utxo.amounts}`);
+      this.log(`Asset: ${utxo.assets}`);
+      this.log(`Commitment: ${utxo._commitment}`);
+      this.log(`Index: ${utxo.index}`);
     }
     this.log("----------------");
   }
