@@ -1,78 +1,105 @@
 import { Command, Flags } from "@oclif/core";
 import * as fs from "fs";
+import {
+  CustomLoader,
+  isValidBase58SecretKey,
+  isValidURL,
+} from "../../utils/utils";
+import { PublicKey } from "@solana/web3.js";
 
 class ConfigCommand extends Command {
   static description = "Update the configuration values";
 
   static flags = {
     rpcUrl: Flags.string({
+      char: "r",
       description: "Solana RPC URL",
     }),
     relayerUrl: Flags.string({
+      char: "l",
       description: "Relayer URL",
     }),
     secretKey: Flags.string({
-      description: "Secret key in string format",
+      char: "s",
+      description: "solana keypair secretkey in base58 string format",
     }),
     relayerRecipient: Flags.string({
+      char: "u",
       description: "Relayer recipient",
     }),
     lookupTable: Flags.string({
+      char: "t",
       description: "Look-up table",
     }),
-    payer: Flags.string({
-      description: "Payer secret key",
-    }),
   };
+
+  protected finally(_: Error | undefined): Promise<any> {
+    process.exit();
+  }
 
   static examples = [
     "$ light config --rpcUrl https://solana-api.example.com",
     "$ light config --relayerUrl https://relayer.example.com",
-    "$ light config --secretKey 0123456789abcdef",
+    "$ light config --secretKey your solana-keypair-secretkey-in-base58-string-format",
     "$ light config --relayerRecipient <recipient_address>",
     "$ light config --lookupTable <lookup_table>",
-    "$ light config --payer <payer_secret_key>",
   ];
 
   async run() {
     const { flags } = await this.parse(ConfigCommand);
 
-    const {
-      rpcUrl,
-      relayerUrl,
-      secretKey,
-      relayerRecipient,
-      lookupTable,
-      payer,
-    } = flags;
+    const { rpcUrl, relayerUrl, secretKey, relayerRecipient, lookupTable } =
+      flags;
+
+    const loader = new CustomLoader("Updating configuration...");
+    loader.start();
 
     try {
       const config = JSON.parse(fs.readFileSync("config.json", "utf-8"));
 
       if (rpcUrl) {
-        config.rpcUrl = rpcUrl;
+        if (isValidURL(rpcUrl)) {
+          config.rpcUrl = rpcUrl;
+        } else {
+          this.error(`\nInvalid URL format`);
+        }
       }
       if (relayerUrl) {
-        config.relayerUrl = relayerUrl;
+        if (isValidURL(relayerUrl)) {
+          config.relayerUrl = relayerUrl;
+        } else {
+          this.error(`\nInvalid URL format`);
+        }
       }
       if (secretKey) {
-        config.secretKey = secretKey;
+        if (isValidBase58SecretKey(secretKey)) {
+          config.secretKey = secretKey;
+        } else {
+          this.error(`\nInvalid solana keypair base58 string format`);
+        }
       }
       if (relayerRecipient) {
-        config.relayerRecipient = relayerRecipient;
+        if (new PublicKey(relayerRecipient)) {
+          config.relayerRecipient = relayerRecipient;
+        } else {
+          this.error(`\nInvalid publickey format`);
+        }
       }
       if (lookupTable) {
-        config.lookUpTable = lookupTable;
-      }
-      if (payer) {
-        config.payer = payer;
+        if (new PublicKey(lookupTable)) {
+          config.lookUpTable = lookupTable;
+        } else {
+          this.error(`\nInvalid publickey format`);
+        }
       }
 
       fs.writeFileSync("config.json", JSON.stringify(config, null, 2));
-
-      this.log("Configuration values updated successfully");
+      this.log("\nConfiguration values updated successfully");
+      loader.stop();
     } catch (err) {
-      this.error(`Failed to update configuration values: ${err}`);
+      loader.stop();
+
+      this.error(`\nFailed to update configuration values: ${err}`);
     }
   }
 }
