@@ -1,13 +1,17 @@
 import { Command, Flags } from "@oclif/core";
-import { getUser } from "../../utils";
+import {
+  CustomLoader,
+  generateSolanaTransactionURL,
+  getUser,
+} from "../../utils/utils";
 import { PublicKey } from "@solana/web3.js";
-import { User } from "light-sdk";
+import { User } from "@lightprotocol/zk.js";
 
 class UnshieldCommand extends Command {
   static description = "Unshield tokens for a user";
 
-  static examples: Command.Example[] = [
-    "$ light unshield --token USDC --amountSpl 1000000 --recipienSpl <address>",
+  static examples = [
+    "$ light unshield --token USDC --amountSpl 1000000 --recipientSpl <address>",
   ];
 
   static flags = {
@@ -16,15 +20,15 @@ class UnshieldCommand extends Command {
       required: true,
     }),
     recipientSpl: Flags.string({
-      description: "The recipient SPL address",
+      description: "The SPL recipient shielded publickey",
     }),
     recipientSol: Flags.string({
-      description: "The recipient SOL address",
+      description: "The SOL recipient shielded/encryption publickey",
     }),
-    amountSpl: Flags.integer({
+    amountSpl: Flags.string({
       description: "The amount of token to unshield (SPL)",
     }),
-    amountSol: Flags.integer({
+    amountSol: Flags.string({
       description: "The amount of token to unshield (SOL)",
     }),
     minimumLamports: Flags.boolean({
@@ -33,6 +37,10 @@ class UnshieldCommand extends Command {
       default: false,
     }),
   };
+
+  protected finally(_: Error | undefined): Promise<any> {
+    process.exit();
+  }
 
   async run() {
     const { flags } = await this.parse(UnshieldCommand);
@@ -46,26 +54,33 @@ class UnshieldCommand extends Command {
       minimumLamports,
     } = flags;
 
+    const loader = new CustomLoader("Performing token unshield...");
+
+    loader.start();
+
     try {
       const user: User = await getUser();
 
       const response = await user.unshield({
         token,
-        recipientSpl: recipientSpl
-          ? new PublicKey(recipientSpl)
-          : PublicKey.default,
-        recipientSol: recipientSol
-          ? new PublicKey(recipientSol)
-          : PublicKey.default,
-        publicAmountSpl: amountSpl,
-        publicAmountSol: amountSol,
+        recipientSpl: recipientSpl ? new PublicKey(recipientSpl) : undefined,
+        recipientSol: recipientSol ? new PublicKey(recipientSol) : undefined,
+        publicAmountSpl: amountSpl ? Number(amountSpl) : undefined,
+        publicAmountSol: amountSol ? Number(amountSol) : undefined,
         minimumLamports,
       });
 
-      this.log(`Successfully unshielded ${token}`);
-      this.log("transaction hash ======>", response.txHash);
+      this.log(
+        `Successfully unshielded ${
+          token.toLowerCase() === "sol" ? amountSol : amountSpl
+        } ${token}`
+      );
+      this.log(generateSolanaTransactionURL("tx", response.txHash, "custom"));
+      loader.stop();
     } catch (error) {
-      this.error(`Unshielding tokens failed: ${error}`);
+      loader.stop();
+
+      this.error(`\nToken unshield failed: ${error}`);
     }
   }
 }
