@@ -772,7 +772,6 @@ export class TransactionParameters implements transactionParameters {
           TransactionParameters.getVerifierConfig(verifierIdl).in,
       });
     }
-
     if (addOutUtxos) {
       outputUtxos = createOutUtxos({
         publicMint: tokenCtx.mint,
@@ -1124,10 +1123,7 @@ export class TransactionParameters implements transactionParameters {
         ...recipientSpl,
         ...this.accounts.recipientSol.toBytes(),
         ...this.relayer.accounts.relayerPubkey.toBytes(),
-        ...new BN(this.relayer.getRelayerFee(this.ataCreationFee)).toArray(
-          "le",
-          8,
-        ),
+        ...this.relayer.getRelayerFee(this.ataCreationFee).toArray("le", 8),
         ...this.encryptedUtxos,
       ]);
 
